Added files for p95v3012b1.

+++ prime95/TortureDlg.h
@@ -1,5 +1,5 @@
 //
-// Copyright 1995-2021 Mersenne Research, Inc.  All rights reserved
+// Copyright 1995-2023 Mersenne Research, Inc.  All rights reserved
 //
 
 #pragma once
@@ -25,36 +25,36 @@ protected:
 public:
 	UINT	m_cores;
 	BOOL	m_hyperthreading;	
-	int m_torture_type;
-	int m_minfft;
-	int m_maxfft;
-	BOOL m_in_place;
-	int m_memory;
-	int m_timefft;
-	int m_avx512;
-	int m_fma3;
-	int m_avx;
-	int m_sse2;
+	int	m_torture_type;
+	int	m_minfft;
+	int	m_maxfft;
+	BOOL	m_in_place;
+	float	m_memory;		// in GB
+	int	m_timefft;
+	int	m_avx512;
+	int	m_fma3;
+	int	m_avx;
+	int	m_sse2;
 	CStatic	c_cores_text;
 	CEdit	c_cores;
 	CButton c_hyperthreading;
 	CButton c_L3_cache;
 	CButton c_L4_cache;
 	CStatic c_minfft_text;
-	CEdit c_minfft;
+	CEdit	c_minfft;
 	CStatic c_maxfft_text;
-	CEdit c_maxfft;
+	CEdit	c_maxfft;
 	CButton c_in_place_fft;
 	CStatic c_memory_text;
-	CEdit c_memory;
+	CEdit	c_memory;
 	CStatic c_timefft_text;
-	CEdit c_timefft;
+	CEdit	c_timefft;
 	CButton c_avx512;
 	CButton c_fma3;
 	CButton c_avx;
 	CButton c_sse2;
 
-	int	m_blendmemory;
+	int	m_blendmemory;			// Default blend memory (in MB)
 
 	afx_msg void OnEnKillfocusCores();
 	afx_msg void OnBnClickedL2Cache();
